fix: change the disguised .webp file to a true .png file
The flower-branch.png was not a valid .png file,
and instead was a .webp file with the extension just renamed to .png.
I noticed it when running optimizations and Oxipng complaining about an invalid file.
When I looked more into the properties of the file,
the structure was of that of the WebP file.

This change makes the size bigger (even after optimizations)
but also makes it not hide the true type of the file.